fix MySQL out of range error

+++ core/includes/updates/200-pr7.php
@@ -45,7 +45,7 @@ try {
         `id` int(11) NOT NULL AUTO_INCREMENT,
         `token` varchar(13) NOT NULL,
         `user_id` int(11) NOT NULL,
-        `discord_user_id` INT(18) NOT NULL,
+        `discord_user_id` bigint(18) NOT NULL,
         PRIMARY KEY (`id`)
         ) ENGINE=$db_engine DEFAULT CHARSET=$db_charset");
 } catch (Exception $e) {
